Add files via upload
This project is in deep learning, and the goal is to try to classify game playing images using CNN (Convolutional Neural Networks). All images are 224 X 224 X 3 in jpg format.  There are 7624 training images. The train, test and validation directories are partitioned into 53 sub directories , one for each of the 53 types of cards. asets.